Add more clarity for user on how to install
When I was trying the utility, I find difficulty to use the scripts. I though the bin files are standalone and can be used as it is even without using the installation. So, I am trying to add some more hints to users who are going to use the utility and not an expert in python